itoa_fast.inl.h已经真正实现, jmc_itoa_r10_fast()和jmc_itoa_r10_fast_ex()已经可用, "%d"的转换速度提升不少; 彻底移除fuint64_t的定义和使用; 其他的一些小调整.

+++ src/jimic/util/utils.h
@@ -66,7 +66,7 @@ jmc_llabs_ex(int64_t *val)
 #if 1
     jmc_int64_t *i64;
     i64 = (jmc_int64_t *)&val;
-    if (i64->i32.negative != 0)
+    if (i64->s64.negative != 0)
         *val = - *val;
 #else
     if (*val < 0)
